integration_2022_wk50

See merge request oai/openairinterface5g!1869

!1663 UL-MIMO working with 2 layers
!1705 RNTI to UE ID in PDCP, SDAP, GTPU, ...
!1818 Move /targets/user to /executables
!1835 Improve NR SRS configuration
!1837 Add Configuration File Naming Style Guide
!1852 Get data from SRB 1 when generating Msg4
!1857 Use F1AP to transfer SRB1 messages in monolithic mode
!1858 fix NR_LDPC_NUM_BN_GROUPS_BG2_R15
!1862 Fix for msg3 scheduling
!1814 Initial web server implementation
!1847 Move 5G RFsim tests from RAN-CI-develop to RAN-Container-Parent
timeout for RFsim

+++ openair2/RRC/NR_UE/rrc_defs.h
@@ -73,7 +73,7 @@ typedef enum {
 #define MAX_UE_NR_CAPABILITY_SIZE 2048
 typedef struct OAI_NR_UECapability_s {
   uint8_t sdu[MAX_UE_NR_CAPABILITY_SIZE];
-  uint8_t sdu_size;
+  uint16_t sdu_size;
   NR_UE_NR_Capability_t *UE_NR_Capability;
 } OAI_NR_UECapability_t;
 
@@ -129,7 +129,7 @@ typedef struct NR_UE_RRC_INST_s {
     uint8_t                        MBMS_flag;
     OAI_NR_UECapability_t          *UECap;
     uint8_t 			   *UECapability;
-    uint8_t                        UECapability_size;
+    uint16_t                       UECapability_size;
 
     RA_trigger_t                   ra_trigger;
     BIT_STRING_t                   requested_SI_List;
